adding filtered function for users
