+++ controllers/user.ts
@@ -1,8 +1,9 @@
-import { type ICustomRequest, type IUserRequest, type IUserResponse, Status } from '../dto'
+import { type ICustomRequest, type IUserRequest, Status } from '../dto'
 import { type Request, type Response } from 'express'
 import { User } from '../models'
 import { createFilter, createUserInstanceService, createUserService, errorMessageHandler, getUserbyIdService } from '../services'
 import { userRepository } from '../repositories'
+import { getUserResponseFiltered } from '../services/user'
 
 const createUser = async (req: ICustomRequest, res: Response): Promise<Response> => {
   try {
@@ -45,17 +46,7 @@ const getAllUsers = async (req: ICustomRequest, res: Response): Promise<Response
 
     const users = await userRepository.find({ relations: { role: true } })
 
-    const usersFiltered = users.map(user => {
-      const { password, role, ...userResponse } = user
-
-      const result: IUserResponse = {
-        ...userResponse,
-        roleId: role.id
-      }
-
-      return result
-    })
-    return res.json(usersFiltered)
+    return res.status(200).json(getUserResponseFiltered(users))
   } catch (error: any) {
     return res.status(500).json(errorMessageHandler(error, 'Error in get all users'))
   }
